requirements.txt
cffi==1.16.0
cryptography==42.0.4
grpcio==1.62.0
grpcio-tools==1.62.0
jwt==1.3.1
protobuf==4.25.3
pycparser==2.21
python-dotenv==1.0.1
